+++ Changelog.md
@@ -16,6 +16,7 @@ Version 1.0 (Tue Aug 09 2011 05:05:41 +0100)
 * Shows the results straight away if there only one when searching Vgmdb
 * The cover art can be dragged to finder or any other place 
 * Pref for how cover art is dragged
+* Debug statement do not appear in release version. 
 
 ### Bug fixes ###
 * Fixed a bug when vgmdb returns no results.